requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
albumentations>=1.3.0
tensorboard>=2.14.0
pycocotools>=2.0.7
fiftyone>=0.21.0  # Optional, for dataset exploration